executer: throw a run error if stack not empty at end of program
TEST: verify that a run error is thrown upon the end of execution if the
stack is not empty (this is a bug since all entries of the stack should
be popped by the time the end code is executed).

Normally this should not occur, but it is a good idea during development
to check in case there is a bug somewhere.  The simulate this error, the
normal compile is not used, instead a single constant number instruction
is added to the program (which will push to the stack during execution).
The program is then run with the expectation that the exception will
occur.

A new run error structure was added with a single message for now (there
will need to be a line number eventually).  At the end of program
exception, the stack is checked (with new stack empty executer access
function), and if not empty, a run error is thrown (no error message
was needed to make the test pass).