project: modified unit tests for complete coverage
Previously the error class destructors were reported as not being
covered by the tests.  This was thought to be due to a problem with the
coverage tools.  The problem is actually because there are two compiler
generated virtual destructors, a static one and a dynamic one.  All
throws were using the static destructors, so these were coverage.

To cover the dynamic destructors, a test was added to simply allocate
a error with the new operator using a unique pointer, which will call
the dynamic destructor when it goes out of scope.

In the constant number other unit test, a test error input support
function was implemented to eliminate duplication.  This function
insured that an error was thrown, and if not returned false, which the
caller checked to make sure was not returned.  However, all tests were
designed to throw an error, so a false was never returned, which showed
up as not being covered.  This function was only used one place, so it
was removed and replaced with the separate check throws as and checks of
the error contents pattern used for the rest of the unit tests.

Added the script used for running the coverage tests and producing the
coverage report.  This script is installed in the build directory ready
to be run (but only when coverage is enabled).

A number of spelling errors among the unit tests were also corrected.