ibc: started implementing standalone command line compiler
TEST: build program and verify usage message when no arguments.

A testing infrastructure needed to be implemented to run a program and
test its output.  This was accomplished using a CMake script that runs
the test.  This script first runs the program with the command line
arguments to test and collects the result of the program into a variable
and sends the output and error streams to an output file.  It checks
that the result matches the expected result, and compares (using the
built-in CMake compare files command) the output file with an expected
output file.

The new command line program was put into its own directory and CMake
build file.  It is linked with the library.  Because the name of this
command is the same as the library name (ultimately the library will be
prefixed with 'lib', but CMake sees these as having the same target
name), the target name was given a '-bin' suffix, and given its
desired executable by setting its output name property.  The test files
were put into a test sub-directory.

The desired usage message was put into an expected usage file in the
test directory.  To make the test pass, the main function simply outputs
the usage message and returns a 1 code (failure).
output