ibc: compile and run a basic source file
TEST: verify that when a simple basic source file name is specified on
the command line, that the program is run, and produces the correct
output.

Just enough code was added to the main function to make this test pass,
including creating a file stream of the input file, creating a program
unit instance, compiling the source and running it.  No error checks
were added since these were not required to make the program run.

OTHER: the run test CMake script was renamed for consistency (camel
casing is not being used for file names), and this script was also
modified to delete the output file upon a passing test (to remove
clutter from the build directories).