ibc: report error if source file cannot be opened
TEST: verify that when the source file cannot be opened, that an error
is reported.

This test gives the command a file that does not exist.  A check was
added to report an error and fail when the file is not open.  Since the
add test statements in the CMake build file were nearly identical, the
duplication was replaced with a single function and three calls to this
function.