ibc: make sure program compiles before running
TEST: verify that is the source file contains errors, that the errors
are reported and the program is not run.

The main function was modified to check for a failed call to compile the
source and report that the program contains errors and fail.

Because the error message contains the file name, the run test CMake
script was modified to set the working directory of the program to the
directory with the BASIC test source files so that just the file name
can be specified in the argument.  This required changing the output
and error file location to the binary (build) directory.  And a new
binary directory parameter was added to be passed to the run test
script.