ibc: also recreate when the also recreate option is specified
TEST: verify when the optional '-r' option is the second argument, that
the program is recreated first, and then executed.

This test uses the simple BASIC source program used for the simple test.
The main function was modified to check the number of arguments - if two
then the file name is the first argument, otherwise it is the second
argument.  To make this test pass, an also recreate flag is set if there
are three arguments (it wasn't necessary to actually check what the
first argument was).