ibc: validate optional argument specified
TEST: verify that if the optional argument is not '-r' then an error
message should be output with the usage message and it should fail.

The main function was modified to put the command line arguments into a
standard vector of standard strings so that easy string comparisons can
be made.  Checks were added if two arguments were specified, the second
is checked to make sure that it is a the '-r' option.  Since the usage
message is now needed twice, the output of it was moved to a new usage
function.

OTHER: it was discovered the wrong program name was output in the usage
message - this was corrected.