windows: corrected problems causing issues on windows
For some reason, there is a problem with the standard header files in
MinGW that cause compile errors with some in line functions.  This was
corrected by added the __NO_INLINE__ definition compile option (only
on Windows).

There was a crash occurring during initialization.  It appears that the
maps in the command code class were not initialized when the code
instances went to add to them from their constructors.  This was
corrected by replacing the static maps with state functions that return
references to the maps, with each function containing the local static
map.  This insures that the map is constructed when it is needed.

There was another crash occurring during initialization where it
appeared that the vectors in code class were added to during the
construction of the code instances, but then got reset.  Again, these
static vectors were changed to static member functions returning
references to the vectors with the functions containing the local static
vectors to insure they are constructed when needed.

These static members did not appear to cause a problem on Linux.

There is a difference in the format output for double precision values
with exponents between Linux and Windows where one and two digit
exponents output two digits on Linux but three on Windows.  The affected
two of the unit tests, which were modified to avoid the issue (the
number of exponents in the test values were set to three).

The CMake build scripts were modified to build the ibc executable in the
main build directory, not in the ibc sub-directory.  There is a problem
on Windows where the executable cannot find and load the ibc library
that is build in the main directory.

windows: changed to build ibc executable in main build directory

Otherwise, the ibc executable is unable to find and load the ibc
library.