ibc: cleaned up and refactored the code.
The ibc main function was becoming large and unrulely, so it was
refactored into two classes and many smaller function.  The first class
is for processing the command lines arguments, and the second class is
for compiling, recreating and executing the program.  For now all the
code remains in the single source file.