windows: moved ibc files to new directory due to conflict
While the changes to make the ibc executable from the main CMake build
script was the solution on Windows so that the executable can find and
load the library, on Linux the name of the executable conflicted with
the directory (on Windows the executable has an '.exe' extension so
there was no conflict).

Therefore, all the ibc files were moved to a new directory to avoid this
conflict on Linux.