compiler: simplified calling expression compiler (refactor)
A pattern was seen developing each time an expression needed to be
compiled where an expression compiler class instance was created,
passing the reference to the compiler instance and calling the
expression compiler function operator function - a somewhat clunky
call.

The expression compiler class is really a sub-class of the compiler
class, so a new compile expression was added to the compiler class
(which now does this clunky call, and it is now only in this one place).
All other calls were replaced with a call to the new compile expression
function, and this simplifies the look of the expression unit tests,
which created the expression compiler instance on a separate statement
(now removed).