negate: add double negate operator code to the code line
TEST: verify that the double negate code is added to the code line
after the double constant code (and operand).

The two earlier negate unit tests were modified to verify that the
constant codes (and operands) are added to the beginning of the code
line, so this covers verifying the constant code and the code size.  The
new test then only needs to verify that the double negate code has been
added.

The use of the require operand macro was no longer appropriate for the
negate unit tests because it is checking for a code size of two, but
with the negate operator, the code size should be three.  A new require
code size operand macro was added that also takes an expected code  size
argument along with the expected data type and number string.  So as not
to add duplicated code, the existing require operand macro was reduced
to calling the new macro with two for the expected code size.  This
prevented having to modify all of the existing uses of this macro where
a code size of two is still expected.

To make this test pass, the compile negation function was modified to
add a double negate code, then return the data type returned by the
compile numeric operand function (because the negate operator returns
the same type as its operand).  This required the double negate code
instance, so one was added.  A null pointer was used for the recreate
and execute function arguments as these were not required to make the
test pass (but were required to make the code compile).