negate: add integer negate operator to the code line
TEST: verify that the integer negate code is added to the code line
after the integer constant code (and operand).

To make this test pass, a check was added to the compile negation
function to add a double code if the compile numeric operand function
returned a double data type, otherwise an integer negate code is added.
This change required an integer negate instance, which for now was given
null pointers for the recreate and execute functions.