negate: corrected to return correct error column
TEST: verify an error is thrown when nothing is after the negate
operator (an "expected numeric expression" error should be thrown).

The change to make this test pass went in a completely unexpected
direction.  This test initially just checked that a compile error would
be throw, which was expected to fail since the compile negation function
was not checking for a null data type from the compile numeric operand
function.  However, the test passed, which meant no new functionality
was needed, and the test was not covering anything the other tests were
not already coverage.  The rest of the test, that checked the error
message, column and length was added.  (All except the column test was
also passing.)

What was occurring was that when the compile negation function called
the compile numeric operand function, it threw the error.  And it turned
out that a null data type would never be returned - either the compile
compile negation function would be called or and error would be thrown.
So the compile negation function will not see a null data type,
therefore it would not need a check for a null data type.

As noted, the column test did fail, which meant something was not
functioning correctly.  This problem was due to the code trying to
call the get column function after peeking for a character at the end
of the stream - once this occurs, the current position (column) can no
longer be read because the input stream is in an error condition.

To correct this problem, the compiler peek next character function was
modified to read the position (column) before peeking at the next
character.  This was the same thing that the get keyword function was
doing before peeking at the next character, therefore, this was removed
and the keyword column member was renamed to just column.  The get
keyword column access function was removed and the get column access
function was changed to return the value of the column member.

The reading of the column was also removed from constant number compiler
parse input function loop as this was essentially doing the same thing.
The column member was removed and its get column access function was
changed to call the compiler get column.  There changes made the column
valid for the error and the test pass.

OTHER: all of the constructor initializers that were initializing
members to constants, were replaced with member initializers in the
class declaration (a C++11 feature) - only two classes were affected.