negate: recreate a negate operator with negative integer constant
TEST: compile a print statement with a negate of a negative integer
constant, and verify that it is recreated correctly.

A new operators source file was created to hold that code for operator
functions.  The double and integer negate code instances were moved
from the expression compiler source file to this new source (replaced
with external declarations).

A negate recreate function was implemented that creates a local string
containing the negate operator character.  This string is swapped with
the top string on the recreator stack.  And the string that was on top
of the stack is appended to the top string on the stack.  The pointer
to this function was put into the negate integer code instance.

A swap top function was added to the recreator class to swap a string
with the string on top of the stack, which was then used in the prepend
keyword.

OTHER: changed to C++11 initialization syntax for all standard input
string stream initializers.