negate: recreate a negate operator with a positive integer constant
TEST: compile an expression with a negation of a positive integer
constant, and verify that is is recreated correctly.

A space needs to be added between the negate operator and an integer
constant when it doesn't begin with a minus sign (in other words, a
digit character).  Otherwise it will be interpreted as negative integer
constant with no negation.  This could be a problem as it would be for
an exponential expression with the exponent operator has higher
precedence and without the space, the meaning of the expression would be
different.

A check was added to the negate recreate function where if the string
operand begins with digit character, a space is appended to the top
string on the stack before the operand (constant in this case) is
appended.