negate: recreate with a positive double constant starting with a period
TEST: compile an expression with a negation of a positive double
constant starting with a decimal point, and verify that it is recreated
correctly.

Double constants can also start with a decimal point in addition to a
digit character, so a space also needs to be added otherwise if will be
treated as a negative double constant instead of a negate operator with
a positive double constant when the expression is compiled again.