expression: implemented initial expected data type mechanism
The compile expression function(s) have an expected data type argument,
but the functionality was not implemented.  This was a place holder as
it was expected it would be needed so that a caller could request a
specific data type (or no data type as for example the PRINT command,
which will take any data type expression).

This broke the TDD process in that code was added that wasn't be tested
(though this didn't show up in the coverage test because coverage can't
detect an unused argument).  It was only added so that the unit tests
would not have to be modified when the mechanism was finally
implemented.

It was realized that the compile numeric operand should not be throwing
an error if no valid [numeric] operand (for now just a number constant)
is found.  The reason is that here a string expression could be valid
(for example for a comparison where the result is a number).  It should
be the upper most caller that should determine if this is an error (or
that a string expression should be attempted).

So to correct this issue, an initial expected data type mechanism was
implemented.  The compile numeric operand was modified to just return
a null data type if no valid constant is found.  The compile negation
function was modified to detect a null data type from compile numeric
operand and throw the "expected numeric expression" error (with out
this, the negate unit tests fail).  A new compile numeric expression
function was added, which calls the next lower function (for now the
compile numeric operand), and if there is an expected data type and a
null data type is returned, the same error is thrown.  The main compile
function was modified to call this function and to pass on the expected
data type argument.

These changes caused several of the column tests to fail because the
column was now not obtained at the correct time.  This was corrected by
adding a new column member to the expression compiler implementation
class, and it set just before calling the compile numeric operand
function (compile negation) or the constant number compiler.  This makes
the column available if an error if thrown.

Finally, since the expected data type must now have an actual value,
the expression unit tests for checking an error when there is nothing in
the input stream needed be modified to request a numeric (double)
expression (otherwise a null data type is returned instead of an error
being thrown.