negate: implemented integer negate operator execution
TEST: compile a statement with negation of an integer constant, and
verify that it executes correctly.

To make this test pass, an execute negate integer function was added
that simply sets the integer on top of the stack to the negation of that
integer.

OTHER: It was noticed that the names of the compile, recreate and
execution functions were not names as per the coding guidelines for this
project, specifically should have been camel casing, not snake casing,
and should have been in the form with the verb at the beginning of the
function name.  All of these were renamed.