exponential: parse both operands and operator of expression
TEST: verify that both operands of the exponential binary operator and
the operator is parsed.

The test initially fails because the exponential operator stops the
compilation of the expression (and only one operand is compiled).

To get this test to pass, a new compile exponential function was added
to the expression compiler.  The compile numeric expression function was
changed to call this new function instead of the compile numeric operand
function.  The new function calls the compile numeric operand function,
then gets the next character (which doesn't need to be checked to make
the program pass), compiles the second operand, and returns the data
type of the first operand (which is necessary to keep all the other
tests passing).