exponential: make sure corrected operator character is present
TEST: verify that parsing stops when an unknown character is found after
an operand.

The letter 'a' character was used as this character will never be a
valid operator.  This test expects that after calling the compile
expression function that the next character is this 'a' character.

This test was initially failing; the whole string was parsed because
the compile exponential function was allowing any character as the
operator character.

To make this test pass, a check was added for the exponential character
'^' and only when this character is present does the second operand get
compiled.  (It is up to the caller to determine if this non-exponential
character is valid.)