exponential: allow white space after the operator
TEST: verify that white space is allowed after the operator character.

This test initially failed because the space caused parsing to be
stopped when compiling the second operand.  The test was made to pass be
skipping white space after consuming the operator character before
compiling the second operand.

OTHER: refactored out the code into new is operator character function
that skips the white space before the operator, checks for the operator
character, and upon finding the operator character, gets the operator
character from the input stream, and skips white space after the
operator.