exponential: make sure there is operand before operator
TEST: verify that if there is an exponential operator at the beginning
of the input stream, that parsing stops at the operator character.

This test initially failed (the second operand was compiled even though
there was not first operand) because the compile exponential function
was not checking for a null data type (no operand) after calling the
compile numeric operand for the first operand.  A check for a null data
type was added to make this test pass.