exponential: throw an error if the second operand is bad
TEST: verify that if there is a problem with the second operand, that
an error is thrown.

The test input stream contains a second exponential character, which is
not a valid operand.  This test initially failed because the compile
exponential function was not checking the returned data type from the
second compiler numeric operand function to a null data type.  To make
this test pass, this check was added.

OTHER: this was the third location for the "expected numeric expression"
error, so to remove code duplication, a new expected numeric expression
error class was add, which derives from compile error.  A test was added
to cover the dynamic destructor of this new error class.