exponential: added operator code to program code
TEST: verify that with two integer operands, that an integer exponential
operator code is added to the end of the program code.

It was not required to also check if the two constant operand codes were
also added to the code line as there are already tests for this (at
least for one constant).  By checking the code length (for five) should
give enough confidence that the two constant integer codes are there
along with the operator code.  The proper constant codes will be
indirectly check when the recreate and execute tests are added later
(once the underlying code is implemented).

To get this test to pass, an integer exponential code instance was added
and this code is added to the program code in the compile exponential
function after checking the data type of the right hand side operand.