exponential: add double code to program code
TEST: verify that with two double operands, that a double exponential
operator code is added to the end of the program code and that a double
data type is returned.

to get this test to pass, a double exponential code instance was added
and this code is added to the program code if the data type of the left
side operand is double (enough to make this and the previous test pass).

OTHER: added a check for an integer data type return to the previous
test.