exponential: add right hand side integer operator code to program code
TEST: verify that with a left hand side double and right hand side
integer operands, a right hand side integer operator code is added to
the end of the program code.

To get this test to pass, a right hand side integer exponential code
instance was added.  Since the logic is growing, the statements to
determine with code to select was moved into a new function.  This
function takes the left and right hand side data types and  returns a
new operator info structure containing a reference to the code and the
result data type of the operator.

In the compile exponential function, the first operand data type was
renaming to left hand side data type, and after calling the new select
function with the two data types and the code added to the program,
the left hand side data type is set to the operand result data type
because this is the variable used to return the data type.