exponential: add left hand side integer operator code to program code
TEST: verify that with a left hand side integer and right hand side
double operands, a left hand side integer operator code is added to
the end of the program code.

To get this test to pass, a left hand side integer exponential code
instance was added.  The select function was modified to check the right
side data type when the left side data type is not a double.