operators: added new class to hold all operand types of operators
The numeric operator codes class holds the four operators for each
combination of operand data types.  The select function was moved to
this class.

A new operator code template class was added where the template
parameter is the operand types (double/double, integer/double,
double/integer, and double/double).  The four exponential operator code
instances were changed to one for a unique operator code class for each
operand type.  This method forces the correct code instances to be given
to the numeric operator codes class constructor correctly.  (The IBCP on
the other hand contained quite a bit of code to verify that the codes
were setup up correctly in the table, but this method does the checking
automatically during compilation.)

The "lhs" and "rhs" terminology was replaced with the actual types for
each side, which is a bit more easier to understand by looking at the
code.