exponential: allow multiple exponential operators in expression
TEST: verify that multiple exponential operators are parsed.

This test was initially failing because the component exponential
function only parsed one exponential operator with its operands and
returned.

To make this test pass, the if statement checking for the exponential
operator was changed to a while so that it will continue parsing
exponential operators.

It wasn't necessary to check if the program code was assembled
correctly - previous unit tests checked the individual operations and
this changes did not affect that part of the code.  This will be checked
later with the recreation and execution tests.