exponential: recreate an expression with two integer operands
TEST: verify that an exponential expression with two integer operands
is recreated correctly.

Note that spaces are added around the operator regardless of whether
they were there in the original expression.

A recreate exponential function was implemented, which creates a local
string with the exponential operator (and a space) and appends the
string of the right hand side operand that is on top of the recreator
stack, which is then popped from the stack (via a new recreator pop
function).  A space is appended to the string on top of the stack (the
left hand side operand), and finally the local string with the operator
and right hand operand is appended to the top string.  A pointer to this
function replaced the null pointer of rhe integer/integer exponential
operator code.