exponential: recreate expressions with other data types
TESTS: verify that exponential expressions with two double operands,
one integer and one double operand, and one double and one integer
operand.

These unit tests were add one at a time.  To make each pass, the null
pointer in the associated codes' recreate function was replaced with
the pointer to the recreate exponential function pointer.

It is not necessary to test the recreation of an expression with
multiple exponential operators - no need code needs to be added to make
this test pass, so the test is unnecessary.  Verification that the
program code for this is compiled correctly will done in the execute
tests.