exponential: handle negative exponents with two integer operands
TEST: verify that when the exponent is negative, that the result is zero
(negative exponents produce fractions of 1 over the number raised to the
exponent, which in the integer domain is always zero).

To make this test pass, a check was added if the exponent (right side
operand) is negative to just set the result to zero.

There are a number of other conditions that will need to be checked for,
which means that this function will continue to grow.  To stem this
growth, a new power integer/integer function operator function class was
added to do the work.  Its constructor receives the x and y operands
(and stores them in local members), and it's function operator function
with no arguments does the work, which first checks if the exponent is
negative (and returns zero), otherwise calls private calculate power
member function that uses multiplication to calculate the result.