const-num: changed from state classes to state member functions
The state class hierarchy used to keep track of the state while parsing
numerical constant was replaced with state member functions, which are
dispatched using member dereference operator.  The base state class
pointer member was changed to a pointer to a member function.

This is a cleaner design since it doesn't require the abstract base,
a bunch of derived virtual classes or the static derived class
instances for each state class.  Also not needed are the several access
functions in the implementation class as the member functions have
direct access to the member variables (a reference argument to the
implementation class instance is not needed).