exponential: added check for zero to negative power (divide by zero)
TEST: verify that when a 0 is raised to a negative exponent that an
error is thrown.

Raising zero to a negative exponent is the same as 1 divided by zero
raise to a positive exponent, which is zero, hence dividing by zero, an
error.

Initially this test did not pass because no error was thrown (probably
resulted in infinity or not-a-number).  A check was added if the number
being raised is a zero to throw a run error (with just a blank string
made the test pass).