exponential: added message and offset to error
TEST: verify a divide by zero error and the correct offset (within the
program where the error occurred).

An offset member was added to the run error class.  This is the offset
within the program that the error occurred.  For the exponential
operator, it is the location of the exponential operator code.  To make
the code compile, all throws of a run error needed an offset argument
that was added to the run error constructor.  To make this test pass,
the offset argument was simply set to two.