end: added correct offset of end code for stack not empty errors
TEST: verify the correct offset (of the end code) when throwing a value
stack not empty at end of program error.

This test initially failed because a dummy zero argument was added to
the throw of this error.  To make the test pass, a new current offset
access function was added to the executer class that calculates the
current offset be subtracting the base code pointer from the current
program counter minus another one since the program counter was already
incremented past the end code.