exponential: added correct offset to divide by zero error
TEST: verify a divide by zero error has the correct offset when the
error occurs at a different offset.

Note: The previously test was checking for an offset of two, which was
hard-coded in the throw of the error, however, the two was not actually
the correct offset, which should have been four.  This was corrected in
the previous test (the throw was changed to have an offset of zero, see
below).

This implementation was problematic and there were two alternatives in
how to implement this change.  The issue is that the divide by zero
error is detected and thrown in the power integer/integer class, which
does not have access to the executer instance from which to get the
current offset.

One solution would be to pass in the executer instance through the
constructor of this class.  The other solution was to catch the error
in the execute exponential integer/integer function, add the offset, and
then rethrow the error.  Both of these are not ideal, but the later was
chosen so as not to penalize passing in another argument to the power
class that unless there is an error will never be used.