program: convert an offset to a line index
TEST: verify that a program offset in a run error is converted back to
the line index containing the offset.

Not seen in this commit was a test with a single line where the line
index expected was zero (the first line).  The initial line index
function implemented in the program unit class simply returned a zero to
make the test pass.  Since this only works with an error in the first
line, the test was modified where the error will occur in the second
line (line index of one).

To get this test to pass, the line index function was fully implemented,
which uses the standard find if function with a find offset lambda
expression that sees if the offset is within the start offset and the
start offset plus size minus one.  The line index is then calculated by
the distance between the resulting iterator and the start of the line
info vector, which is returned.  If the offset is not found (like for
the offset of the automatic END), the size of the line info vector is
returned (indicating beyond the end of the program).

For complete testing, though not required for coverage, a series of
boundary offset tests (begin and end offset for each line, and offsets
beyond the end) were also added to verify that the lambda expression is
correct and that offset beyond the end word as desired.