program: added code to convert the error offset to a column
TEST: verify that the column of the error can be obtained from the
offset of the error and the line index.

To obtain the column for the offset of the operator, the program line
first needs to be recreated (since the original line is no longer
present, and may not match the recreated line anyway).  And while the
line is being created, it needs to mark where the error is.

This was accomplished by adding an error offset member to the recreator
class.  An error offset argument was added to the constructor.  An at
error offset flag member was also added that the main recreate loop sets
when at the offset of the error.  The recreate exponential function
calls new mark error start recreate member function, which will append a
special [control] character to the string on top of the stack if the at
error offset flag is set.

A begin iterator was added to the program reader class that is used by a
new current offset access function that returns the current offset of
the reader (the difference between the begin and current iterator).
This function is used by the recreator to determine if it is at the
error offset.

The program unit recreate line was given an optional error offset
argument (defaults to -1, or the highest unsigned integer value, which
will not match any offset).  The error offset is passed to the recreator
constructor.

A new error info access function was added to the program unit class
taking a line index and error offset.  It calls the recreate line to
created the selected line.  The returned program line string will then
contain the special error start marker character.  New private member
function extract error info is called with the program line string,
which searches for the special character.  A new error info class
containing just the column for now is returned at setting the column.

The recreate exponential function was rearranged a bit to separate the
spaces around the operator, the operator and the call to mark the error
start.

OTHER: Some additional constant modifiers were added to several
functions including the dictionary get and recreate get constant number
operand functions, and new constant number dictionary access constant
function was added.  In order for this to work, the recreator program
reader instance was made mutable (allowing it to be modified with a
constant recreator instance).