program: added length to program run error info
TEST: verify the length of the error.

For now an error length of one is sufficient, which is the length of
the exponential operator character.  A length member was added to the
error info structure, which is simply set to one.  This will be changed
when there is an error requiring more than a length of one.