program: added line string to program run error info
TEST: verify the program line string recreated from the program for the
line with the run error.

The text of the line with the error needs to be returned in the error
info structure so that it can be presented to the user along with a
pointer to the location of the error.

The recreated line contains the start error marker character, which
needs to be removed from the program line string before it is returned
in the error info.  The character is erased using the column (position)
the was previously obtained.

The program unit extract error info function was only using the error
info structure (setting its members) that this function was extracted as
a new constructor of the error info structure.  The error info function
was changed to create (construct) the error info instance from the
program line.