program: cleaned up run error handling and interface
Renamed the run function to the execute function, which was made
private.  This function creates the executer instance, runs the program,
and checks for a stack error.  Run errors are thrown out of this
function.  Implemented a new public run function that calls the private
execute function, catches any run error and converts it to a program
error (containing the line number, column, length and line text).

The error info structure was changed to the program error class (and
also includes the error message).  Two constructors are provided, one
for errors at the end of the program where just the run error instance
is passed (for just the message, the line text string will be empty),
and one that is also passed the line number and recreated line.  The
program unit error info function was no longer used and was removed, and
the line index function was made private.

Changed the executer current offset return value from integer to an
unsigned integer to be compatible with the standard size type.  Made the
same change to the offset member of the run error class.

Changed tests to use the new program error class and check it contents
(line number, column, length, line text and error message) instead of
the offset that was in the run error class (and no longer needs the line
index function, which was made private).  There were now some
unnecessary duplicated tests, so these were removed (along with the
boundary tests, which will be okay since the program unit line index
function was tested - any further changes will require new tests).